Updated all product components, also added checkout page
Updated all the products pages, so now when adding or deleting items from the database the HTML will add new products into a grid layout modularly and reflect the information when clicking on it. The only thing that needs to be fixed it the image that corresponds with each product. I also added a checkout page with no CSS, the routing is correct and functions as needed.